+++ core/types/liquidity.go
@@ -16,6 +16,7 @@ import (
 	"errors"
 	"fmt"
 	"strings"
+	"time"
 
 	"code.vegaprotocol.io/vega/libs/num"
 	proto "code.vegaprotocol.io/vega/protos/vega"
@@ -42,6 +43,55 @@ const (
 	// margin check fails, then they will be cancelled without any penalties.
 	LiquidityProvisionStatusPending LiquidityProvisionStatus = proto.LiquidityProvision_STATUS_PENDING
 )
+
+type LiquiditySLAParams struct {
+	PriceRange                      num.Decimal
+	CommitmentMinTimeFraction       num.Decimal
+	ProvidersFeeCalculationTimeStep time.Duration
+	PerformanceHysteresisEpochs     uint64
+	SlaCompetitionFactor            num.Decimal
+}
+
+func (l LiquiditySLAParams) IntoProto() *proto.LiquiditySLAParameters {
+	return &proto.LiquiditySLAParameters{
+		PriceRange:                      l.PriceRange.String(),
+		CommitmentMinTimeFraction:       l.CommitmentMinTimeFraction.String(),
+		ProvidersFeeCalculationTimeStep: int64(l.ProvidersFeeCalculationTimeStep.Seconds()),
+		PerformanceHysteresisEpochs:     l.PerformanceHysteresisEpochs,
+		SlaCompetitionFactor:            l.SlaCompetitionFactor.String(),
+	}
+}
+
+func LiquiditySLAParamsFromProto(l *proto.LiquiditySLAParameters) *LiquiditySLAParams {
+	return &LiquiditySLAParams{
+		PriceRange:                      num.MustDecimalFromString(l.PriceRange),
+		CommitmentMinTimeFraction:       num.MustDecimalFromString(l.CommitmentMinTimeFraction),
+		ProvidersFeeCalculationTimeStep: time.Second * time.Duration(l.ProvidersFeeCalculationTimeStep),
+		PerformanceHysteresisEpochs:     l.PerformanceHysteresisEpochs,
+		SlaCompetitionFactor:            num.MustDecimalFromString(l.SlaCompetitionFactor),
+	}
+}
+
+func (l LiquiditySLAParams) String() string {
+	return fmt.Sprintf(
+		"priceRange(%s) commitmentMinTimeFraction(%s) providersFeeCalculationTimeStep(%v) performanceHysteresisEpochs(%v) slaCompetitionFactor(%s)",
+		l.PriceRange.String(),
+		l.CommitmentMinTimeFraction.String(),
+		l.ProvidersFeeCalculationTimeStep,
+		l.PerformanceHysteresisEpochs,
+		l.SlaCompetitionFactor.String(),
+	)
+}
+
+func (l LiquiditySLAParams) DeepClone() *LiquiditySLAParams {
+	return &LiquiditySLAParams{
+		PriceRange:                      l.PriceRange,
+		CommitmentMinTimeFraction:       l.CommitmentMinTimeFraction,
+		ProvidersFeeCalculationTimeStep: l.ProvidersFeeCalculationTimeStep,
+		PerformanceHysteresisEpochs:     l.PerformanceHysteresisEpochs,
+		SlaCompetitionFactor:            l.SlaCompetitionFactor,
+	}
+}
 
 type TargetStakeParameters struct {
 	TimeWindow    int64